rename : sweetAlert 경로 변경
