+++ src/pages/profile_edit_detail/profile_edit_detail.js
@@ -1,7 +1,7 @@
 import '@/layout/footer';
 import '@/styles/pages/profile_detail.scss';
 import { getNode, getNodes, setStorage } from '@/library/index';
-import { sweetConfirm, sweetBasic, sweetError } from '@/layout/sweetAlert';
+import { sweetConfirm, sweetBasic, sweetError } from '@/components/sweetAlert';
 import { getMyProfile, updateRecord } from '@/api/getRecords';
 import gsap from 'gsap';
